// src/types/userProfile.d.ts
export interface UserProfile {
  id: number;
  firstName: string;
  lastName: string;
  role: string;
  username: string;
  email: string;
  phoneNumber: string;
  profileImageUrl: string;
  eventsApplied: number;
  eventsHeld: number;
  currentEvents: number;
  linkedInUrl: string;
  birthday: string;
  biography: string;
}